content: news: 2025-10-31: News articles must use H3 or lower
Because "News" is H1, and new title is H2, article body must be H3 or
lower. This means that you can only use `###` or `####`.

Currently, "News" isn't H1 actually. And some says that H1 should be
news title and I kida agree. So it might be changed but for now it's
what we've got.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>